코드 간결화
