+++ src/main/resources/static/js/notmain.js
@@ -14,14 +14,18 @@ function stealingjson(){
             alert("예외 발생 : 코드 똑바로 만드시오.");
         });
 }
-
-function ihatework(){   // *.json 파일 제작용 메소드, URL 따라가면서 전부다 @삭제
+// 서버 켠 후 한번만 눌러줄것, 챠트에 필요한 데이타들 메모리에 올리는 작업 실행
+// todo: 지금은 수동으로 하지만 나중에는 json파일 업데이트 될 때 마다 또는 서버 켜지면 자동으로 실행되도록 자동화 해야함.
+function loadData(){
     $.ajax({
-        url:'/ihatework',
+        url:'/statistics/loaddata',
         method:'get',
-        success:function (json) {
-            console.log(json)
-            alert('console.log를 확인하세요')
+        success:function (args) {
+            if(args===true){
+                alert('정상적으로 메모리에 적재가 완료되었읍니다.');
+            }else {
+                alert('예외발생. 콘솔 읽어보셈');
+            }
         }
     });
 }
