+++ src/main/resources/templates/notmain.html
@@ -6,9 +6,9 @@
 </head>
 <body>
     <h3>핼로우, 월드!</h3>
-    <button type="button" onclick="stealingjson()">Stealing JSON</button>
+    <button type="button" onclick="stealingjson()">누르지 마시오</button>
     <button type="button"><a href="/view">WorldWide_Statistics</a></button>
-    <button type="button" onclick="ihatework()">노가다 싫어 메소드</button>
+    <button type="button" onclick="loadData()">메모리에 데이타 올리기</button>
